separate Auth and User routes
